examples/basic-python/requirements.txt
powerlobster-webhook>=1.0.0
python-dotenv>=1.0.0
